Fix random_search
